requirements.txt
fastapi
uvicorn
google-adk
google-cloud-aiplatform
vertexai
requests
numpy
python-dotenv
